linebot-project/requirements.txt
flask==3.0.2
line-bot-sdk==3.9.0
python-dotenv==1.0.1
requests==2.31.0
googlemaps==4.10.0
firebase-admin==6.4.0
google-generativeai==0.3.2
python-dateutil==2.8.2
gunicorn==21.2.0